Organize queries by module
