+++ packages/stargate/src/modules/auth/queries.spec.ts
@@ -6,9 +6,9 @@ import { BaseAccount } from "cosmjs-types/cosmos/auth/v1beta1/auth";
 import { Any } from "cosmjs-types/google/protobuf/any";
 import Long from "long";
 
-import { nonExistentAddress, pendingWithoutSimapp, simapp, unused, validator } from "../testutils.spec";
-import { AuthExtension, setupAuthExtension } from "./auth";
-import { QueryClient } from "./queryclient";
+import { QueryClient } from "../../queryclient";
+import { nonExistentAddress, pendingWithoutSimapp, simapp, unused, validator } from "../../testutils.spec";
+import { AuthExtension, setupAuthExtension } from "./queries";
 
 async function makeClientWithAuth(
   rpcUrl: string,
