+++ packages/stargate/src/modules/index.ts
@@ -1,0 +1,9 @@
+export { AuthExtension, setupAuthExtension } from "./auth/queries";
+export { BankExtension, setupBankExtension } from "./bank/queries";
+export { DistributionExtension, setupDistributionExtension } from "./distribution/queries";
+export { GovExtension, GovParamsType, GovProposalId, setupGovExtension } from "./gov/queries";
+export { IbcExtension, setupIbcExtension } from "./ibc/queries";
+export { MintExtension, MintParams, setupMintExtension } from "./mint/queries";
+export { setupSlashingExtension, SlashingExtension } from "./slashing/queries";
+export { setupStakingExtension, StakingExtension } from "./staking/queries";
+export { setupTxExtension, TxExtension } from "./tx/queries";
