+++ packages/stargate/src/modules/mint/queries.ts
@@ -3,9 +3,7 @@ import { assert } from "@cosmjs/utils";
 import { Params } from "cosmjs-types/cosmos/mint/v1beta1/mint";
 import { QueryClientImpl } from "cosmjs-types/cosmos/mint/v1beta1/query";
 
-import { createProtobufRpcClient } from "../";
-import { QueryClient } from "./queryclient";
-import { decodeCosmosSdkDecFromProto } from "./utils";
+import { createProtobufRpcClient, decodeCosmosSdkDecFromProto, QueryClient } from "../../queryclient";
 
 /**
  * Like Params from "cosmjs-types/cosmos/mint/v1beta1/mint"
