+++ packages/stargate/src/modules/slashing/queries.ts
@@ -6,8 +6,7 @@ import {
 } from "cosmjs-types/cosmos/slashing/v1beta1/query";
 import { QueryClientImpl } from "cosmjs-types/cosmos/slashing/v1beta1/query";
 
-import { QueryClient } from "./queryclient";
-import { createPagination, createProtobufRpcClient } from "./utils";
+import { createPagination, createProtobufRpcClient, QueryClient } from "../../queryclient";
 
 export interface SlashingExtension {
   readonly slashing: {
