+++ packages/stargate/src/modules/staking/queries.spec.ts
@@ -4,9 +4,10 @@ import { Tendermint34Client } from "@cosmjs/tendermint-rpc";
 import { sleep } from "@cosmjs/utils";
 import { MsgDelegate, MsgUndelegate } from "cosmjs-types/cosmos/staking/v1beta1/tx";
 
-import { MsgDelegateEncodeObject, MsgUndelegateEncodeObject } from "../encodeobjects";
-import { SigningStargateClient } from "../signingstargateclient";
-import { assertIsDeliverTxSuccess } from "../stargateclient";
+import { MsgDelegateEncodeObject, MsgUndelegateEncodeObject } from "../../encodeobjects";
+import { QueryClient } from "../../queryclient";
+import { SigningStargateClient } from "../../signingstargateclient";
+import { assertIsDeliverTxSuccess } from "../../stargateclient";
 import {
   defaultSigningClientOptions,
   faucet,
@@ -14,9 +15,8 @@ import {
   simapp,
   simappEnabled,
   validator,
-} from "../testutils.spec";
-import { QueryClient } from "./queryclient";
-import { setupStakingExtension, StakingExtension } from "./staking";
+} from "../../testutils.spec";
+import { setupStakingExtension, StakingExtension } from "./queries";
 
 async function makeClientWithStaking(
   rpcUrl: string,
