--- packages/stargate/src/queries/index.ts
@@ -1,21 +1,0 @@
-// Base symbols
-
-export { QueryClient } from "./queryclient";
-
-// Extensions
-
-export { AuthExtension, setupAuthExtension } from "./auth";
-export { BankExtension, setupBankExtension } from "./bank";
-export { DistributionExtension, setupDistributionExtension } from "./distribution";
-export { GovExtension, GovParamsType, GovProposalId, setupGovExtension } from "./gov";
-export { IbcExtension, setupIbcExtension } from "./ibc";
-export { MintExtension, MintParams, setupMintExtension } from "./mint";
-export { setupSlashingExtension, SlashingExtension } from "./slashing";
-export { setupStakingExtension, StakingExtension } from "./staking";
-export { setupTxExtension, TxExtension } from "./tx";
-export {
-  createPagination,
-  createProtobufRpcClient,
-  decodeCosmosSdkDecFromProto,
-  ProtobufRpcClient,
-} from "./utils";
