+++ packages/stargate/src/queryclient/index.ts
@@ -1,0 +1,8 @@
+export { QueryClient } from "./queryclient";
+export {
+  createPagination,
+  createProtobufRpcClient,
+  decodeCosmosSdkDecFromProto,
+  longify,
+  ProtobufRpcClient,
+} from "./utils";
