+++ packages/stargate/src/queryclient/utils.ts
